Add better step output for ECS steps
Using a sidecar container to report back to the step function allows us to get better control of the output.

Right now we're only supplying a link to the logs, but this can easily be extended to supply more information.

The actual sidecar image should probably be separated into another repo, but right now I'm keeping it here for easier testing.