Fix python dependencies causing failed apply
This fixes an issue where the module couldn't be applied because the
host that applies the terraform doesn't have gcc or g++.

This is fixed by explicitly installing all python dependencies.

Preferably, the lambda would be packaged as a container image, but I
haven't found a good enough solution for this yet.